move painter out of mjpeg.dart
